fix(overlay): isolate shortcut formatter
